rename semconv->attributes
